+++ triad-sample/app/src/main/java/com/nhaarman/triad/sample/MainComponent.java
@@ -1,12 +1,8 @@
 package com.nhaarman.triad.sample;
 
-import flow.Flow;
 import org.jetbrains.annotations.NotNull;
 
 public class MainComponent {
-
-  @NotNull
-  private final Flow mFlow;
 
   @NotNull
   private final NoteRepository mNoteRepository;
@@ -17,16 +13,10 @@ public class MainComponent {
   @NotNull
   private final NoteValidator mNoteValidator;
 
-  public MainComponent(@NotNull final Flow flow) {
-    mFlow = flow;
+  public MainComponent() {
     mNoteRepository = new MemoryNoteRepository();
     mNoteCreator = new NoteCreator(mNoteRepository);
     mNoteValidator = new NoteValidator();
-  }
-
-  @NotNull
-  public Flow flow() {
-    return mFlow;
   }
 
   @NotNull