+++ triad-sample/app/src/main/java/com/nhaarman/triad/sample/notes/NotesScreen.java
@@ -4,7 +4,6 @@ import com.nhaarman.triad.sample.MainComponent;
 import com.nhaarman.triad.sample.R;
 import com.nhaarman.triad.sample.notes.noteslist.NotesListPresenter;
 import com.nhaarman.triad.screen.Screen;
-import dagger.Component;
 import org.jetbrains.annotations.NotNull;
 
 public class NotesScreen extends Screen<NotesPresenter, NotesContainer, MainComponent> {
@@ -17,16 +16,7 @@ public class NotesScreen extends Screen<NotesPresenter, NotesContainer, MainComp
   @NotNull
   @Override
   protected NotesPresenter createPresenter(@NotNull final MainComponent mainComponent) {
-    return Dagger_NotesScreen_NotesComponent.builder().mainComponent(mainComponent).build().notesPresenter();
+    NotesListPresenter notesListPresenter = new NotesListPresenter(mainComponent.noteRepository());
+    return new NotesPresenter(notesListPresenter);
   }
-
-
-  @NotesScope
-  @Component(dependencies = MainComponent.class)
-  interface NotesComponent {
-
-    NotesPresenter notesPresenter();
-
-  }
-
 }